Gjør mapping mellom dataklasser i servicer; organiserer dataklasser slik at dataklasser fra k9-selvbetjening-oppslag ligger i gateway og dataklasser brukt mot pleiepenge-soknad ligger i 'egen' fil
