+++ src/main/kotlin/no/nav/helse/arbeidsgiver/Arbeidsgivere.kt
@@ -3,3 +3,8 @@ package no.nav.helse.arbeidsgiver
 data class Arbeidsgivere (
     val organisasjoner: List<Organisasjon>
 )
+
+class Organisasjon (
+    val organisasjonsnummer: String,
+    val navn: String?
+)
