--- src/main/kotlin/no/nav/helse/arbeidsgiver/Organisasjon.kt
@@ -1,6 +1,0 @@
-package no.nav.helse.arbeidsgiver
-
-class Organisasjon (
-    val organisasjonsnummer: String,
-    val navn: String?
-)
