+++ src/main/kotlin/no/nav/helse/barn/BarnService.kt
@@ -14,14 +14,21 @@ class BarnService(
     internal suspend fun hentNaaverendeBarn(
         ident: String,
         callId: CallId
-    ) =
-        try {
-            barnGateway.hentBarn(
-                personIdent = ident,
-                callId = callId
-            )
+    ) = try {
+        barnGateway.hentBarn(
+            personIdent = ident,
+            callId = callId
+            ).map { it.tilBarn() }
         } catch (cause: Throwable) {
             logger.error("Feil ved henting av barn, returnerer en tom liste", cause)
             emptyList<Barn>()
         }
+
+    private fun BarnGateway.BarnOppslagDTO.tilBarn() = Barn(
+        fodselsdato = fødselsdato,
+        fornavn = fornavn,
+        mellomnavn = mellomnavn,
+        etternavn = etternavn,
+        aktoerId = aktør_id
+    )
 }
